Use more roq features
Move JS into built bundle to minimize server trips
Use mvnpm rather than CDNs to avoid extra requests per page load
Add supports or LLMS specification for easier agentic discovery
Remove jquery usage, converting JS to vanilla
