--- src/main/resources/web/app/index.js
@@ -1,1 +1,0 @@
-import './main.scss';
