+++ templates/layouts/base.html
@@ -1,31 +1,27 @@
+---
+twitter: WildFlyAS
+---
+
 <!DOCTYPE html>
 <html>
 
 <head>
   <title>{page.title}</title>
-  <!--<script id="adobe_dtm" src="//www.redhat.com/dtm.js" type="text/javascript"></script>-->
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta name="description" content="{page.description ?: site.description}">
   <link rel="me" href="https://fosstodon.org/@wildflyas" />
-  <link rel="shortcut icon" type="image/png" href="{site.url('/favicon.ico')}" >
-  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.1.0/css/all.css" integrity="sha384-lKuwvrZot6UHsBSfcMvOkWwlCMgc0TaWr+30HWe3a4ltaBwTZhyTEggF5tJv8tbt" crossorigin="anonymous">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/styles/atom-one-dark.min.css">
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/highlight.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/protobuf.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/dockerfile.min.js"></script>
-  {! <link rel="stylesheet" href="{site.url('/assets/css/main.css')}" /> !}
   <!-- Global site tag (gtag.js) - Google Analytics -->
   <script async src="https://www.googletagmanager.com/gtag/js?id=UA-40221748-1"></script>
   {|
   <script>
-    window.dataLayer = window.dataLayer || [];
-    function gtag(){dataLayer.push(arguments);}
-    gtag('js', new Date());
-
-    gtag('config', 'UA-40221748-1');
+      window.dataLayer = window.dataLayer || [];
+      function gtag(){dataLayer.push(arguments);}
+      gtag('js', new Date());
+      gtag('config', 'UA-40221748-1');
   </script>
   |}
+  <link rel="shortcut icon" type="image/png" href="{site.url('/favicon.ico')}"/>
 
   {#if page.data.containsKey('external_link')}
   <meta http-equiv="refresh" content="0; {page.data.external_link}">
@@ -58,16 +54,11 @@
 
   {#include partials/project-footer.html /}
   {#include partials/commonhaus-footer.html /}
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>
-  <script type="text/javascript" src="{site.url('/assets/javascript/mobile-nav.js')}"></script>
   <script type="text/javascript">
     if (("undefined" !== typeof _satellite) && ("function" === typeof _satellite.pageBottom)) {
         _satellite.pageBottom();
     }
-    hljs.highlightAll();
   </script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/clipboard.js/2.0.11/clipboard.min.js"></script>
-  <script type="text/javascript" src="{site.url('/assets/javascript/copy.js')}"></script>
 
 </body>
