+++ templates/layouts/guides.html
@@ -60,5 +60,3 @@ layout: base
         {/for}
     </div>
 </div>
-
-<script src="{site.url('/assets/javascript/guides-search.js')}"></script>
